PAI1/requirements.txt
fpdf
python-docx